# 🧠 Natural Memory Triggers with Performance-Aware Architecture