… .claude/settings.local.json
… claude-hooks/README-NATURAL-TRIGGERS.md
… claude-hooks/config.json
… claude-hooks/debug-pattern-test.js
… claude-hooks/simple-test.js
… claude-hooks/utilities/git-analyzer.js
… claude-hooks/utilities/performance-manager.js
… claude-hooks/utilities/tiered-conversation-monitor.js