@doobidoo 2025-09-27
/gemini review

Please review this comprehensive Natural Memory Triggers implementation. This is a major enhancement that adds intelligent mid-conversation memory awareness to Claude Code with performance optimization.

Key areas for review:
- Architecture design and component integration
- Performance optimization and tiered processing
- Pattern detection accuracy and learning mechanisms
- Test coverage and quality assurance
- Configuration management and user experience
- Code quality and maintainability

The implementation introduces sophisticated natural language processing for detecting when users need memory context, with configurable performance profiles to balance intelligence with responsiveness.