@doobidoo 2025-09-27
## ✅ All Gemini Code Assist Recommendations Addressed

I have systematically fixed every issue identified in the Gemini Code Assist review:

### **HIGH Priority Fixes**
- ✅ **Hardcoded Performance Profiles** - Modified `PerformanceManager.getProfileBudget()` to prioritize configuration over hardcoded defaults
- ✅ **Flaky Timing Test** - Replaced exact timing assertions with functional behavior testing to eliminate CI failures

### **MEDIUM Priority Fixes**
- ✅ **Duplicate Configuration Keys** - Removed duplicate `sensitivity` and `adaptiveLearning` from naturalTriggers section 
- ✅ **Division-by-Zero Protection** - Added safety checks in `AdaptivePatternDetector.checkSemanticMatch()` and `TieredConversationMonitor.calculateSemanticShift()`
- ✅ **Inconsistent CLI Documentation** - Standardized all command examples to use full path `claude-hooks/memory-mode-controller.js`
- ✅ **Code Quality** - Removed unused variables, extracted magic numbers to constants
- ✅ **Running Average Bug** - Fixed feedback counting logic in `updateAcceptanceRate()` function

### **Verification Results**
- 🧪 **18/18 tests passing** (100% success rate)
- ⚙️ **Configuration-based profiles working** correctly
- 🛡️ **Division-by-zero protections validated**
- 📊 **Feedback counting logic verified**

All fixes have been thoroughly tested and are ready for re-review.

/gemini review